backend/requirements.txt
flask==3.0.3
openai-whisper
deep-translator==1.11.4
gtts==2.5.1
flask-cors==4.0.1
moviepy==1.0.3